Reorganize models and adjust code accordingly
