--- lib/Model/User/Preference.php
@@ -1,80 +1,0 @@
-<?php
-/**
- * @author Ilja Neumann <ineumann@owncloud.com>
- *
- * @copyright Copyright (c) 2018, ownCloud GmbH
- * @license GPL-2.0
- *
- * This program is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License as published by the Free
- * Software Foundation; either version 2 of the License, or (at your option)
- * any later version.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for
- * more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program; if not, write to the Free Software Foundation, Inc.,
- * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
- *
- */
-namespace OCA\DataExporter\Model\User;
-
-class Preference {
-	/** @var string */
-	private $appId;
-	/** @var string */
-	private $configKey;
-	/** @var string */
-	private $configValue;
-
-	/**
-	 * @return string
-	 */
-	public function getAppId(): string {
-		return $this->appId;
-	}
-
-	/**
-	 * @param string $appId
-	 * @return Preference
-	 */
-	public function setAppId(string $appId): Preference {
-		$this->appId = $appId;
-		return $this;
-	}
-
-	/**
-	 * @return string
-	 */
-	public function getConfigKey(): string {
-		return $this->configKey;
-	}
-
-	/**
-	 * @param string $configKey
-	 * @return Preference
-	 */
-	public function setConfigKey(string $configKey): Preference {
-		$this->configKey = $configKey;
-		return $this;
-	}
-
-	/**
-	 * @return string
-	 */
-	public function getConfigValue(): string {
-		return $this->configValue;
-	}
-
-	/**
-	 * @param string $configValue
-	 * @return Preference
-	 */
-	public function setConfigValue(string $configValue): Preference {
-		$this->configValue = $configValue;
-		return $this;
-	}
-}
