+++ tests/unit/Exporter/MetadataExtractor/FilesExtractorTest.php
@@ -24,7 +24,7 @@ namespace OCA\DataExporter\Tests\Unit\Exporter\MetadataExtractor;
 
 use OCA\DataExporter\Exporter\MetadataExtractor\FilesExtractor;
 use OCA\DataExporter\Utilities\Iterators\Nodes\RecursiveNodeIteratorFactory;
-use OCA\DataExporter\Model\User\File;
+use OCA\DataExporter\Model\UserMetadata\User\File;
 use OCP\Files\Node;
 use OCP\Files\Folder;
 use Test\TestCase;
