+++ tests/unit/Exporter/MetadataExtractor/PreferencesExtractorTest.php
@@ -23,7 +23,7 @@
 namespace OCA\DataExporter\Tests\Unit\Exporter\MetadataExtractor;
 
 use OCA\DataExporter\Exporter\MetadataExtractor\PreferencesExtractor;
-use OCA\DataExporter\Model\User\Preference;
+use OCA\DataExporter\Model\UserMetadata\User\Preference;
 use OCP\IAppConfig;
 use OCP\IConfig;
 use Test\TestCase;
